timeseries_environment_dockerfull/requirements.txt
kafka-python==2.0.2
influxdb-client==1.45.0
torch>=2.2.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
python-dotenv>=1.0.0